Fixed missing handler method on error (WebTestClient) (#250)
The current implementation of `WebTestClientInitializer` only works if the rest endpoint succeeds. In case of an error (for example passing an `String` as `PathVariable` of type `int`) the handler method is not determined correctly and this leads for example to a missing headline in the `auto-section.adoc`.

The reason is that spring webflux `DispatcherHandler` first calls all implementations of `HandlerAdapter` and only if they succeeds the `HandlerResultHandler` implementations (the `WebTestClientInitializer` is one of them) are called (see [`DispatcherHandler.java#L156-L161`](https://github.com/spring-projects/spring-framework/blob/v5.0.2.RELEASE/spring-webflux/src/main/java/org/springframework/web/reactive/DispatcherHandler.java#L156-L161)).

To solve this issue I simply changed the `WebTestClientInitializer` to implement `HandlerAdapter` instead of `HandlerResultHandler`.